requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==1.0.1
google-generativeai==0.3.2
python-multipart==0.0.9
typing-extensions==4.9.0
h11==0.14.0 